requirements.txt
tqdm~=4.66.1
pycryptodome
tqdm
